+++ src/net/tmwa/generalhandler.cpp
@@ -132,8 +132,8 @@ void GeneralHandler::handleMessage(Net::MessageIn &msg)
     {
         case SMSG_CONNECTION_PROBLEM:
         {
-            const int code = msg.readInt8();
-            logger->log("Connection problem: %i", code);
+            const uint8_t code = msg.readUInt8();
+            logger->log("Connection problem: %u", static_cast<int>(code));
 
             switch (code)
             {
